Create 6-psu-server.py
This scrip can confugure servertech pdu (master-slave) for servers where PDU-A is master and PDU-B is slave. same port sall be used for a host in PDU-A and B.

to use this script make two .txt file in same loaction.
hostname.txt and ports.txt

for switches it consider 2 psu per switch, PSU1 on PDU-A and PSU-2 on PDU-B port number being same on both PDU.

after switch give a blank line and than servers.
it considers 6 psu per server. odd PSU(1,2,3) on PDU-A and even PSU(2,4,6) on PDU-b.